sample_vector_embeddings/requirements.txt
openai==1.43.0
python-dotenv==1.0.1
